Fix 1459615 (debian #364408): allow building of synfigstudio on amd64. Patch by Andreas Jochens.
git-svn-id: http://svn.voria.com/code@158 1f10aa63-cdf2-0310-b900-c93c546f37ac

+++ synfig-studio/trunk/src/gtkmm/dockdialog.cpp
@@ -87,7 +87,7 @@ DockDialog::DockDialog():
 	widget_comp_select=new Widget_CompSelect();
 	
 	// Give ourselves an ID that is most likely unique
-	set_id(synfig::UniqueID().get_uid()^reinterpret_cast<int>(this));
+	set_id(synfig::UniqueID().get_uid()^reinterpret_cast<long>(this));
 	
 	set_role(strprintf("dock_dialog_%d",get_id()));
 	GRAB_HINT_DATA(
